iio:st_magn: enable device after trigger
Currently, we enable the device before we enable the device trigger. At high
frequencies, this can cause interrupts that don't yet have a poll function
associated with them and are thus treated as spurious. At high frequencies with
level interrupts, this can even cause an interrupt storm of repeated spurious
interrupts (~100,000 on my Beagleboard with the LSM9DS1 magnetometer). If these
repeat too much, the interrupt will get disabled and the device will stop
functioning.

To prevent these problems, enable the device prior to enabling the device
trigger, and disable the divec prior to disabling the trigger. This means
there's no window of time during which the device creates interrupts but we have
no trigger to answer them.

Signed-off-by: Martin Kelly <martin@martingkelly.com>