iio:st_magn: enable device after trigger
I'd like to get others' opinions on this patch before I submit it. It fixes
an issue I've seen with a driver I'm working on (LSM9DS1 magnetometer
driver, waiting on this before I send a patch). However, I have a few
concerns about the generality of this approach:

(1) Possible data races between st_sensors_set_enable and IRQ handlers,
    since both could run concurrently. Although we can protect this with a
    lock, we'd need to touch each driver, which will be tricky to test.
    Alternatively, we could audit each driver and see if there are any real
    races. The only property touched by st_sensors_set_enable appears to be
    the "enabled" property, which is unlikely to be read/written by any IRQ
    handlers. That said, the fact that the race exists makes me concerned
    about potential mistakes in the future.
(2) If this is a good approach to solve the issue, then we should make the
    same change to the ST accelerometer, gyroscope, and pressure drivers.
    But again, we'd need to re-test every driver, which is tricky.

I'd be happy to hear opinions or suggested alternative approaches. Thanks!

Patch:
--------------------------------------------------------------------------
Currently, we enable the device before we enable the device trigger. At
high frequencies, this can cause interrupts that don't yet have a poll
function associated with them and are thus treated as spurious. At high
frequencies with level interrupts, this can even cause an interrupt storm
of repeated spurious interrupts (~100,000 on my Beagleboard with the
LSM9DS1 magnetometer). If these repeat too much, the interrupt will get
disabled and the device will stop functioning.

To prevent these problems, enable the device prior to enabling the device
trigger, and disable the divec prior to disabling the trigger. This means
there's no window of time during which the device creates interrupts but we
have no trigger to answer them.

Signed-off-by: Martin Kelly <martin@martingkelly.com>